+++ packages/constructor/src/constructors/Avatar/index.ts
@@ -1,0 +1,5 @@
+export * from './Avatar'
+export * from './AvatarDesign'
+export * from './props'
+export * from './types'
+export * from './basicTypes'