+++ packages/d1/src/components/Ui/Avatar/wiki.ts
@@ -1,0 +1,13 @@
+import { WikiStorybook } from '@dxtmisha/wiki'
+import { wiki, wikiDescriptions } from '@dxtmisha/wiki/media'
+
+import { AvatarWikiData } from './wikiData'
+
+export const AvatarWikiStorybook = new WikiStorybook(
+  AvatarWikiData.component,
+  AvatarWikiData.props,
+  AvatarWikiData.defaults,
+  AvatarWikiData.wikiDesign,
+  wiki,
+  wikiDescriptions
+)